Refatora punishment() para módulo functions
